# openapi.yaml
openapi: 3.0.3
info:
  title: FinMate API
  description: |
    API Contract for FinMate, a Personal Finance & Lifestyle Companion.
    Provides endpoints for Auth, Users, Groups, Expenses, Settlements, Notes, Goals, and Import/Export.

    ### Versioning Strategy
    URL-based versioning is used. All paths start with `/api/v1`.

    ### Pagination & Filtering
    Paginated GET endpoints support standard query parameters:
    * `page`: 1-based page index (default: `1`)
    * `limit`: Page size (default: `20`, max: `100`)
    * `cursor`: Cursor string (for cursor-based pagination on rapidly updated lists like Expenses and Audit Logs)
  version: 1.0.0
servers:
  - url: /api/v1
    description: Local Dev Server
paths:
...
  /groups:
    post:
      summary: Create an expense group
      tags:
        - Groups
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  maxLength: 120
                  example: 'Flatmates Goa Trip'
                description:
                  type: string
                  example: 'Sharing expenses for our trip to Goa'
                visibility:
                  type: string
                  enum: [private, invite_only, public_readonly]
                  default: private
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    get:
      summary: List groups current user is a member of
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: isArchived
          in: query
          schema:
            type: boolean
          required: false
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /groups/{id}:
    get:
      summary: Get details of a group
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      summary: Update group metadata or archive group
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - version
              properties:
                name:
                  type: string
                  maxLength: 120
                description:
                  type: string
                visibility:
                  type: string
                  enum: [private, invite_only, public_readonly]
                isArchived:
                  type: boolean
                version:
                  type: integer
                  example: 1
      responses:
        '200':
          description: Group updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /groups/{id}/members:
    post:
      summary: Invite or add a member to the group
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - role
              properties:
                email:
                  type: string
                  format: email
                  example: friend@example.com
                role:
                  type: string
                  enum: [admin, member, viewer]
                  default: member
      responses:
        '201':
          description: Member added/invited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMember'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
    get:
      summary: List all members of the group
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of group members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupMember'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /groups/{id}/members/{memberId}:
    patch:
      summary: Update a group member role or join status
      description: |
        Used by admins to modify member roles, or by the member themselves to accept/decline/leave group.
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [owner, admin, member, viewer]
                joinStatus:
                  type: string
                  enum: [invited, active, left, removed]
      responses:
        '200':
          description: Member updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMember'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Remove a member from the group or leave the group
      tags:
        - Groups
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Member removed successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of elements per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 400
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/auth/register'
            errorCode: 'VAL_INVALID_INPUT'
            message: 'Input validation failed'
            details:
              - field: 'email'
                issue: 'must be a valid email address'
            retryable: false
    UnauthorizedError:
      description: Authentication failed or token missing/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/users/me'
            errorCode: 'AUTH_TOKEN_EXPIRED'
            message: 'JWT access token has expired'
            retryable: false
    ForbiddenError:
      description: Permissions not met for resource action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 403
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/groups/d3b07384-d113-49c3-a558-f54248880bf6'
            errorCode: 'RES_FORBIDDEN'
            message: 'You do not have access to view this group.'
            retryable: false
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/expenses/f81d4fae-7dec-11d0-a765-00a0c91e6bf6'
            errorCode: 'RES_NOT_FOUND'
            message: 'Expense with ID f81d4fae-7dec-11d0-a765-00a0c91e6bf6 does not exist.'
            retryable: false
    ConflictError:
      description: Resource constraint or duplication conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 409
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/auth/register'
            errorCode: 'RES_ALREADY_EXISTS'
            message: 'A user with email user@example.com already exists.'
            retryable: false
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - timestamp
        - path
        - errorCode
        - message
        - retryable
      properties:
        statusCode:
          type: integer
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        errorCode:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            required:
              - field
              - issue
            properties:
              field:
                type: string
              issue:
                type: string
        retryable:
          type: boolean
    PaginatedResponse:
      type: object
      required:
        - data
        - meta
        - links
      properties:
        meta:
          type: object
          required:
            - totalItems
            - itemCount
            - itemsPerPage
            - totalPages
            - currentPage
          properties:
            totalItems:
              type: integer
            itemCount:
              type: integer
            itemsPerPage:
              type: integer
            totalPages:
              type: integer
            currentPage:
              type: integer
        links:
          type: object
          required:
            - first
            - last
          properties:
            first:
              type: string
            previous:
              type: string
              nullable: true
            next:
              type: string
              nullable: true
            last:
              type: string
...
    Group:
      type: object
      required:
        - id
        - name
        - description
        - visibility
        - ownerUserId
        - isArchived
        - version
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        visibility:
          type: string
          enum: [private, invite_only, public_readonly]
        ownerUserId:
          type: string
          format: uuid
        isArchived:
          type: boolean
        version:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GroupMember:
      type: object
      required:
        - id
        - groupId
        - userId
        - role
        - joinStatus
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, admin, member, viewer]
        joinStatus:
          type: string
          enum: [invited, active, left, removed]
        joinedAt:
          type: string
          format: date-time
          nullable: true
        leftAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...